Add files via upload
This is the latest of our pivotal tracker and we are doing for the bus schedule. We are doing 4 user story which is the Bus Schedule Homepage, Add bus schedule information, Update bus information and bus homepage. Choose which part of the user story you want to involve it. Noted that it should be average work to be done by all members. For the add bus schedule information with have 3 points of difficulty, more people are required to do it.